End of Day 8 Commit
Did a Complete overhaul off the Spell System, it is now fully functioning allowing for me to create dozens of unique spells that operate of slight changes to overridable logic.
Hooked the Spell System up to UI and got a Casting Progress Bar intrgrated into the game and working.  The bar fades out when the player is not casting any Spells.