End Of Day 9 Commit
Added the ability to swap Player Allies's postion around, included UI elements that allowed for this and fully coded the functinoality. Along with a toggle visibility function for the ui so it is hidden when not in use.
Also added a StatusEffect Class which will allow certain spells the ability to buff or debuff allies, enemies or even the player.